Retrieve every tag when converting a mesh to and from meshio
Tags travel as point_sets/cell_sets, EasyFEA_to_Gmsh writes gmsh:physical instead of a key
meshio ignores, and reading picks that array rather than letting the zeroed gmsh:geometrical
overwrite it. A name that is not P{i}/L{i}/S{i}/V{i} no longer raises, and an unnamed
reference 0 no longer invents a tag on a mesh that had none.

+++ EasyFEA/FEM/_group_elem.py
@@ -1757,7 +1757,7 @@ class _GroupElem(ABC):
             return self.__dict_elements_tags[tag]
         else:
             print(f"The {tag} tag is unknown")
-            return np.array([])
+            return np.array([], dtype=int)
 
     def Get_Nodes_Tag(self, tag: str) -> _types.IntArray:
         """Returns node associated with the tag."""
@@ -1765,7 +1765,7 @@ class _GroupElem(ABC):
             return self.__dict_nodes_tags[tag]
         else:
             print(f"The {tag} tag is unknown")
-            return np.array([])
+            return np.array([], dtype=int)
 
     def Locates_sol_e(
         self, sol: _types.FloatArray, dof_n: Optional[int] = None, asFeArray=False
